Update ServiceDTO.cs and PetsitterDTO.cs

+++ backend/Petsupport.API2/Dtos/OutDtos/PetsitterDTO.cs
@@ -9,7 +9,6 @@ namespace Petsupport.API2.Dtos.OutDtos
         public string City { get; set; }
         public string Street { get; set; }
         public string Description { get; set; }
-        //TODO: check this
-        public decimal Price { get; set; }
+        public double Price { get; set; }
     }
 }
